Adds a POJ-specific accent mark normalization rule: The ui-rule.
TL only has iu rule, i.e. although by sonority rule i > u, but if u comes
after i, we put the accent mark on u on typographical grounds. For POJ
there is an opinion that the combination ui should follow the same rule.
As it's not in the TL rules, we make it POJ specific.

Input methods that rely on the default TL input type and POJ output type
will have to double-normalize the output POJ syllable to apply the rule,
because convertToPOJSyllable() method does not normalize the original
syllable (by design, so as to preserve the intended accent marks;
normalization should always be an external operation performed upon the
syllable object).